+++ src/view/result/query.ts
@@ -1,3 +1,4 @@
+import * as mysql from "mysql";
 import { MessageType, OperateType, ConfigKey } from "../../common/constants";
 import { Node } from "../../model/interface/node";
 import { ColumnNode } from "../../model/other/columnNode";
@@ -8,12 +9,16 @@ import { DataResponse } from "./queryResponse";
 import { Global } from "../../common/global";
 import { ExportService } from "../../service/export/exportService";
 import { MysqlExportService } from "../../service/export/impl/mysqlExportService";
+import { PageService } from "../../service/page/pageService";
+import { MysqlPageSerivce } from "../../service/page/impl/mysqlPageSerivce";
+import { ConnectionManager } from "../../service/connectionManager";
 
 export class QueryParam<T> {
     /**
      * using in loadColumnList.
      */
     public connection?: Node;
+    public singlePage?: boolean;
     public type: MessageType;
     public res: T;
 }
@@ -21,8 +26,13 @@ export class QueryParam<T> {
 export class QueryPage {
 
     private static exportService: ExportService = new MysqlExportService()
+    private static pageService: PageService = new MysqlPageSerivce()
 
     public static async send(queryParam: QueryParam<any>) {
+
+        if (typeof queryParam.singlePage == 'undefined') {
+            queryParam.singlePage = true;
+        }
 
         switch (queryParam.type) {
             case MessageType.DATA:
@@ -38,6 +48,7 @@ export class QueryPage {
         }
 
         ViewManager.createWebviewPanel({
+            singlePage: queryParam.singlePage,
             splitView: !Global.getConfig(ConfigKey.QUERY_FULL_SCREEN),
             path: "pages/result/index", title: "Query",
             initListener: (webviewPanel) => {
@@ -47,6 +58,13 @@ export class QueryPage {
                 switch (params.type) {
                     case OperateType.execute:
                         QueryUnit.runQuery(params.sql);
+                        break;
+                    case OperateType.next:
+                        const sql = this.pageService.build(params.sql, params.pageNum, params.pageSize)
+                        const connection = await ConnectionManager.getConnection(ConnectionManager.getLastConnectionOption())
+                        QueryUnit.queryPromise(connection, sql).then((rows) => {
+                            QueryPage.send({ type: MessageType.NEXT_PAGE, res: { sql, data: rows } as DataResponse });
+                        })
                         break;
                     case OperateType.export:
                         this.exportService.export(params.sql)