+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.1 (2020/5/6)
+- Highlight active database.
+- Enhance dashboard.
+- Support open table in new panel.
+
+## 1.9.6 (2020/5/1)
+- Fix bugs.
+
 ## 1.9.0 (2020/4/28)
 - Support SSH tunel.
 - Show comment of column、table on tree view.